Refactor ActiveModel::Dirty and ActiveRecord::Dirty
Previously, you could include ActiveRecord::Dirty and the included hook
would check what the class was, and from there include the appropriate
Dirty module. This mean that Mobility was including more modules than it
really had to.

Instead, we're now just assuming that if you include the AR::Dirty
plugin, you're including it in an AR::Base class (if not we raise a
TypeError). This assumption allows us to simplify things and avoid
including extra modules, keeping the class' ancestors clean and more
understandable. Also, I've made sure that all Dirty modules other than
the Attributes one have constants so they are easy to see (as opposed to
anonymous modules.)